Refatorando o código
